+++ src/components/aocs/magnetorquer.cpp
@@ -6,7 +6,7 @@
 #include "magnetorquer.hpp"
 
 #include <interface/log_output/logger.hpp>
-#include <library/math/global_randomization.hpp>
+#include <library/randomization/global_randomization.hpp>
 #include <library/math/matrix_vector.hpp>
 #include <library/math/quaternion.hpp>
