+++ src/components/aocs/magnetorquer.hpp
@@ -9,9 +9,9 @@
 #include <environment/local/local_environment.hpp>
 #include <interface/log_output/loggable.hpp>
 #include <library/math/matrix.hpp>
-#include <library/math/normal_randomization.hpp>
+#include <library/randomization/normal_randomization.hpp>
 #include <library/math/quaternion.hpp>
-#include <library/math/random_walk.hpp>
+#include <library/randomization/random_walk.hpp>
 #include <library/math/vector.hpp>
 
 #include "../base/component.hpp"
