+++ src/components/aocs/star_sensor.cpp
@@ -8,7 +8,7 @@
 #include <environment/global/physical_constants.hpp>
 #include <interface/log_output/log_utility.hpp>
 #include <library/math/constants.hpp>
-#include <library/math/global_randomization.hpp>
+#include <library/randomization/global_randomization.hpp>
 #include <library/math/matrix.hpp>
 #include <string>
