+++ src/components/aocs/star_sensor.hpp
@@ -9,10 +9,10 @@
 #include <dynamics/attitude/attitude.hpp>
 #include <environment/local/local_environment.hpp>
 #include <interface/log_output/loggable.hpp>
-#include <library/math/normal_randomization.hpp>
 #include <library/math/quaternion.hpp>
-#include <library/math/randomize1.hpp>
 #include <library/math/vector.hpp>
+#include <library/randomization/minimal_standard_linear_congruential_generator_with_shuffle.hpp>
+#include <library/randomization/normal_randomization.hpp>
 #include <vector>
 
 #include "../base/component.hpp"
