+++ src/components/aocs/sun_sensor.hpp
@@ -9,7 +9,7 @@
 #include <environment/local/local_celestial_information.hpp>
 #include <environment/local/solar_radiation_pressure_environment.hpp>
 #include <interface/log_output/loggable.hpp>
-#include <library/math/normal_randomization.hpp>
+#include <library/randomization/normal_randomization.hpp>
 #include <library/math/quaternion.hpp>
 #include <library/math/vector.hpp>
