+++ src/components/base/sensor.hpp
@@ -7,8 +7,8 @@
 #define S2E_COMPONENTS_BASE_SENSOR_HPP_
 
 #include <library/math/matrix.hpp>
-#include <library/math/normal_randomization.hpp>
-#include <library/math/random_walk.hpp>
+#include <library/randomization/normal_randomization.hpp>
+#include <library/randomization/random_walk.hpp>
 #include <library/math/vector.hpp>
 
 /**
