+++ src/components/ideal_components/force_generator.hpp
@@ -9,7 +9,7 @@
 #include <components/base/component.hpp>
 #include <dynamics/dynamics.hpp>
 #include <interface/log_output/logger.hpp>
-#include <library/math/normal_randomization.hpp>
+#include <library/randomization/normal_randomization.hpp>
 #include <library/math/vector.hpp>
 
 /*
