+++ src/environment/local/atmosphere.cpp
@@ -6,9 +6,9 @@
 #include "atmosphere.hpp"
 
 #include <interface/log_output/log_utility.hpp>
-#include <library/math/global_randomization.hpp>
-#include <library/math/normal_randomization.hpp>
-#include <library/math/random_walk.hpp>
+#include <library/randomization/global_randomization.hpp>
+#include <library/randomization/normal_randomization.hpp>
+#include <library/randomization/random_walk.hpp>
 #include <library/math/vector.hpp>
 
 using libra::NormalRand;
