+++ src/environment/local/geomagnetic_field.cpp
@@ -8,9 +8,9 @@
 #include <library/external/igrf/igrf.h>
 
 #include <interface/initialize/initialize_file_access.hpp>
-#include <library/math/global_randomization.hpp>
-#include <library/math/normal_randomization.hpp>
-#include <library/math/random_walk.hpp>
+#include <library/randomization/global_randomization.hpp>
+#include <library/randomization/normal_randomization.hpp>
+#include <library/randomization/random_walk.hpp>
 
 using libra::NormalRand;
 using namespace std;
